Change Realm & Firebase to use their xcframeworks, halve build time (#1410)
Uses the binary distribution of Firebase and Realm, rather than compiling it on-device every time. These are gigantic dependencies that we don't really need to see the source for unless we want to.

I effectively copied the `.podspec.json` from the spec repo and updated it to use the downloads which contain the binary frameworks. Maintaining them should be as simple as changing the URL and version numbers for updates. Should that fail, reverting to the default pods still works without any non-Podfile changes.

Clean debug build time, iOS: 310s → 150s. macOS is now at 95s since it has less archs (no watchOS simulator). Tests in CI similarly decrease from 16m to 8m.