Allow entering nabu.casa URLs manually (#1448)
## Summary
Removes the restriction around entering nabu.casa URLs directly. 

## Any other notes
We do not gain much by disallowing it, other than support requests, and the most common case is that the remote UI isn't actually set up or is having SSL issues, which this will validate and check the URL before saving, providing an error message if appropriate.