Fix crash in share extension (#1585)
I can't figure out when this started crashing, maybe a recent Xcode release, but it crashes due to the `init(nibName:bundle:)` initializer not being generated when the `init(coder:)` initializer is marked as unavailable. Strange. It crashes not just on the latest iOS 14 but also in iOS 13, so it's almost certainly a generated code and not runtime issue.