Manual complication updates via notifications (#1615)
Depends on home-assistant/mobile-apps-fcm-push#52. Fixes #1046, though not in the best way. However, this at least leaves it open to doing it better later.

## Summary
Adds support for triggering complication updates indirectly via background notifications to the iOS app, and adds a manual update button to the complication list.

## Screenshots
![Image](https://user-images.githubusercontent.com/74188/116802546-6c2cf980-aac8-11eb-913a-6b9df9153ede.png)

## Link to pull request in Documentation repository
Documentation: home-assistant/companion.home-assistant#511

## Any other notes
This indirect method appears to be a bit slower than the direct method, but seems to mostly work.