Fix tests and Watch App execution in Xcode 12.5 (#1696)
## Summary
Works around Xcode 12.5 / 13.0 (beta 1/2 at least) regressions in target launching.

## Any other notes
When running tests for iOS simulator, we get an error like:

> The bundle identifier for Home Assistant Launcher.app couldn’t be read. No such file or directory: “/Users/zac/Library/Developer/Xcode/DerivedData/HomeAssistant-ahjyjjgpphztaebycwnsuwmrdiho/Build/Products/Debug/Home Assistant Launcher.app”.

The Launcher.app target is set to macOS only, so it's not built for iOS simulator. Building it first then executing tests works, so clearly Xcode is incorrectly trying to analyze the target when it isn't used.

This also happens when building for Catalyst, where it can't find the watchOS app which isn't being compiled since it is iOS only:

> The bundle identifier for HomeAssistant-WatchApp.app couldn’t be read. No such file or directory: “/Users/zac/Library/Developer/Xcode/DerivedData/HomeAssistant-ahjyjjgpphztaebycwnsuwmrdiho/Build/Products/Debug-watchos/HomeAssistant-WatchApp.app”.

Filed this as FB9189357 as well.