Handle confirmable local notifications (#1831)
Requires home-assistant/core#54947. Refs #1799.

## Summary
Advertises as supporting confirm for local pushes and sends confirmation when receiving a local push.

## Any other notes
Confirmable notifications means the app is required to acknowledge receipt of a notification to core, otherwise it will be sent over the normal non-local-push route. This should eliminate a class of potential bugs in the app, where notifications fail to deliver via local push for some reason; however, this means that there's a possibility that notifications may be delayed if such a bug is occurring, so it is hopefully maybe still noticeable/annoying enough to get reports that we can trace back to find the root cause.