Try both internal and external URLs in onboarding (#1954)
A common case appears to be an internal we cannot connect to (outdated IP address or SSL to a local IP) and an external address we are able to connect to (like Nabu Casa). Detect this and switch to talking to just the external if the internal fails connectivity checks.