Fix time sensitive entitlement not being added to app (#1999)
## Summary
Apparently when I did this in #1676, I ended up checking in an "echo" in front of the entitlements change, so it never worked. Sigh.

## Link to pull request in Documentation repository
Documentation: home-assistant/companion.home-assistant#634

## Any other notes
This was only flagged out anyway because of trying to keep Xcode 12 compatibility, which we are long-since-passed, so always include it now.