Install cocoapods plugin in README, update paths
This adds the installation step required for cocoapods-acknowledgements
to the readme (where `pod` is installed via Homebrew rather than
Bundler). The call to `pod install` would previously fail with:

    [!] Your Podfile requires that the plugin `cocoapods-acknowledgements` be installed. Please install it and try installation again.

This also updates the Homebrew to be dynamically determined, allowing
the install instructions to work on Apple Silicon and for any Homebrew
installation location.