Add option to open links in Private browser tab
This currently only applies to Firefox as it is the only browser with
this ability (Focus/Klar are private by default and Safari/Chrome don't
have a way to open a private or incognito tab at present). That said,
the code is written to be extensible should this change later; the new
"Open in Private Tab" UI in settings only shows up when a supported
browser is selected.

Note that we can't force Firefox to open a link in a "non-private" tab;
it intentionally uses the last mode Firefox was in by default when
opening URLs.