Don't automatically consider hardware address for internal URL (#2302)
## Summary
Stops inserting the current network hardware address, since it may be
hardware which is connected to other networks. Allow the user to choose
this later.

## Any other notes
Since we can't directly tell when a hardware address is a valid way to
consider internal, don't automatically include it. This resolves an
issue where we'd think any Wi-Fi connection was internal which is
invalid logic.