Fix incorrectly confirming a local notification even if it fails (#2329)
## Summary
Fixes #2316, but not the underlying problem, by not confirming local
pushes which fail to be added to the notification center.

## Any other notes
The underlying problem is that we cannot attach anything to a
notification in the Local Push extension as of iOS 16.4. This is clearly
a regression, and it doesn't matter where we write it, it fails to
locate it in locations vended by FileManager:

- The temporary directory
- The documents directory
- The shared container for our app group

All of these fail to attach, all in the same way, and all of them:

1. Are able to be read by the process using `Data(contentsOf: …)`
2. Work fine in the Notification Service Extension

This is definitely a regression on Apple's side and frustratingly my
radar has gone ignored. Par for the course, I guess.