Add sahredEventDeviceInfo data into the "tag_scanned" event
sharedEventDeviceInfo contains "sourceDevicePermanentID",  "sourceDeviceName" and "sourceDeviceID" values. 

It is used in other action events (such as event triggered from the Apple Watch). "sourceDeviceName" can be used to send an actionable notification back to the device that was used to scan tag. Currently it's not possible, since these values are not dispatched in the event body.