Dark mode snapshot strategies (#3482)
<!-- Thank you for submitting a Pull Request and helping to improve Home
Assistant. Please complete the following sections to help the processing
and review of your changes. Please do not delete anything from this
template. -->

## Summary
This PR provides a helper method to automatically assert against light
and dark mode snapshots for SwiftUI views.

It's already applied for LocationHistoryListView and Widgets.

Apologies for the long PR. But it should be quite a visual one as most
changes are new dark mode snapshot tests.

## Link to pull request in Documentation repository
@bgoncal you might be able to point me to where I could update
documentation about snapshot tests.