Snapshot preview reuse (#3484)
<!-- Thank you for submitting a Pull Request and helping to improve Home
Assistant. Please complete the following sections to help the processing
and review of your changes. Please do not delete anything from this
template. -->

## Summary
SnapshottablePreviews applied in the LocationHistoryListView and
Widgets.

In LocationHistoryListView, due to how Realm is used and how previews
are implemented under the hood, the injected data is written both on
`init` and `onAppear`.

For Widgets, not all the previews are visible. This is a limitation of
the previews UI.

I also ended up removing the duplicate snapshots that would assert the
behavior when more than the allowed tiles were passed. This can be
restored if preferred.

---------

Co-authored-by: Bruno Pantaleão Gonçalves <5808343+bgoncal@users.noreply.github.com>